[tests] moved tests for commands to test_commands.py
